median class
keda feh error f el median f el graph w mesh mrboot b el main gui "do ur
job ya 7azem :p"

+++ [TEMPLATE] ImageFilters/ImageFilters/quicksort.cs
@@ -4,42 +4,42 @@ using System.Text;
 
 namespace Sort
 {
-	class quicksort
-	{
-		static public int Partition(int[] numbers, int left, int right)
-		{
-			int pivot = numbers[left];
-			while (true)
-			{
-				while (numbers[left] < pivot)
-					left++;
+     class quicksort
+     {
+          static public int Partition(byte[] numbers, int left, int right)
+          {
+               int pivot = numbers[left];
+               while (true)
+               {
+                    while (numbers[left] < pivot)
+                         left++;
 
-				while (numbers[right] > pivot)
-					right--;
+                    while (numbers[right] > pivot)
+                         right--;
 
                     if (left < right)
-                        KthElement.Swap(ref numbers[left], ref numbers[right]);
+                         KthElement.Swap(ref numbers[left], ref numbers[right]);
                     else
                     {
                          return right;
                     }
-			}
-		}
+               }
+          }
 
-		static public void QuickSort_Recursive(int[] arr, int left, int right)
-		{
-			// For Recusrion
-			if (left < right)
-			{
-				int pivot = Partition(arr, left, right);
+          static public void QuickSort_Recursive(byte[] arr, int left, int right)
+          {
+               // For Recusrion
+               if (left < right)
+               {
+                    int pivot = Partition(arr, left, right);
 
-				if (pivot > 1)
-					QuickSort_Recursive(arr, left, pivot - 1);
+                    if (pivot > 1)
+                         QuickSort_Recursive(arr, left, pivot - 1);
 
-				if (pivot + 1 < right)
-					QuickSort_Recursive(arr, pivot + 1, right);
-			}
-		}
-	}
-	}
+                    if (pivot + 1 < right)
+                         QuickSort_Recursive(arr, pivot + 1, right);
+               }
+          }
+     }
+}
